Fewer constraints and image output
Instead of writing a constraint for every triplet of image points, write down a constraint for every pair containing the values of /all/ intermediate points along the line between them. This requires introducing a slack variable for each constraint.

Also, display the output as a multi-colored image with each convex region colored differently.